Add tests for `HtmlListItem.computeIntrinsic` methods
